Update remaining counter information
